requirements.txt
streamlit
Pillow
python-docx
python-pptx
yt-dlp
comtypes; platform_system == "Windows"
